smart-cv/cv-processor/requirement.txt
grpcio-tools
python-dotenv
confluent-kafka[avro,json,protobuf]
minio
scikit-learn
sentence-transformers